make variable magit-hide-diffs unnecessary
Variable `magit-hide-diffs' was used to let-bind
`magit-section-hidden-default' at the appropriate time, which is right
before ("around") creating the section whose subsections should be
hidden.

The latter variable also shouldn't exist.  While that is a different
issue, it is actually related: This commit is the first step required
to eventually remove the latter.

`magit-hide-diffs' was used to explicitly tell the code used to insert
the diff-containing sections, whether that section should be hidden.

However passing this information explicitly is not necessary because we
have implicit knowledge of then it should be hidden: when the section is
only one among other "top" sections.  Or in other words when the mode
isn't `magit-log-mode' or `magit-commit-mode'.

In `magit-wash-raw-diff' we can unconditionally hide the section,
because it is only ever used to insert sections that should be hidden.