magit-file-log: require confirmation to use default file
Previously, when `magit-file-log' was invoked from a file-visiting
buffer, a prefix argument had to be used to select another file.  Now
the current file is merely the default, which the user can confirm by
simply pressing RET.  This is good for two reasons: it is similar to
how other commands (e.g. `kill-buffer') handle this, and the behavior
no longer depends on what kind of buffer is current when the command
is invoked.