magit-file-log: use magit-refresh-log-buffer
In `magit-file-log' use the generic `magit-refresh-log-buffer'
(and pass it `magit-custom-options'), instead of the inferior
`magit-refresh-file-log-buffer'.  The latter was an incomplete
reimplementation of the former "specialized" for "file logs".
In the previous commit we taught the latter the FILE argument;
nothing else was necessary to make it suitable for "file logs".