use existing abstractions to implement cherry
Remove `magit--refresh-cherry-buffer' which reimplemented functionality
already available elsewhere.  `git cherry' can be seen as a `git log'
with less options but an additional twist.  Teach `magit-format-log-line'
and `magit-parse-log-line' the new `cherry' style.  Define the format in
`magit-log-cherry-re'.

Define a washer `magit-insert-cherry-commits' that works like other "log"
washers in that in only washes the list of commits.  Add inserters that
insert header line sections, and add `magit-cherry-insert-sections-hook'
to allow reordering.  This can also be used remove the help text, we
display by default because the old implementation did it too.

Always make the help text less verbose.  Don't add spaces around the
cherry type indicator (+ or -), because that only looked good if the
faces used to display them did set the background color.

Note that "Head" means something else in the cherry buffer than the
status buffer.  This ambiguity is inherited from Git, but we might be
able to do something about it later.  The "Upstream" header still has
to be improved.  I had intended to use `magit-format-tracked-line' for
that purpose but that isn't able to deal with the ambiguity of user
input.  And since that is an issue that also effects other areas I'll
leave it at that for now, and don't attempt to prettify the upstream.