magit-wash-log: when washing cherry output, reverse it first
When `magit-wash-log' washes `git cherry' output it now starts washing
by reversing the git command's output.  This makes the order of cherries
consistent with the order of unpushed/unpulled commits as displayed in
the status buffer.

The order of commits in such "preview" commit lists should probably be
customizable; for now we just make it consistent.