magit-cherry: change argument order
Switch the position of `magit-cherry's UPSTREAM and HEAD arguments.
In interactive use it makes more sense to read the HEAD first, because
that allows defaulting to the tracked branch when later reading UPSTREAM.
Do not just change the order of reading, actually change the order of the
arguments so that interactive and non-interactive use is consistent.

However do not change the order accordingly for function
`magit-refresh-cherry-buffer' and in the value of variable
`magit-refresh-args'.  Here we should instead be consistent with `git
cherry'.  Also changing the order here would complicate matters.

Don't refuse to run when HEAD (the actual git HEAD, not the argument) is
detached.  It is possible to select any rev not just refs as the cherry
head.  And why refuse to run with a detached HEAD if the user is
actually about to select some ref?