add faces magit-cherry-{equivalent,unmatched}
Add the faces `magit-cherry-equivalent' and `magit-cherry-unmatched',
defining them from scratch.  Do not inherit from `magit-diff-{add,del}.
These faces are added exactly because they should not both look the same
as the diff face the corresponds to the same symbol `+' or `-'.  The +┌-
distinction for diffs is distinct from the +┌- distinction for cherries.

Use boring foreground colors that are likely to work with both light and
dark themes.  Do not set the background because that would be overridden
by section highlighting.