add preliminary implementation of magit-format-ref-label
Add a preliminary implementation of function `magit-format-ref-label'
which returns a list whose only element is a propertized ref label; or
nil.  It should actually return just the propertized label; always.

This function is a wrapper around `magit-ref-get-label-color' and
lessens the pain of using that somewhat.  However it also inherits some
of the limitations, including but not limited to, the above.  The
biggest problem is does both filtering and formatting.  Actually fixing
this and other issues would require some refactoring.  I am adding this
kludge instead so that I can land some other things first, that have
been on the waiting list for to long already.