magit-revert-buffer: remove incomplete comment
Remove half truth about why we are not reverting an indirect buffer.
Sure, as the comment said, that would revert the parent buffer; but
the reason we don't want that is relevant: we would end up doing it
twice.  We *could* extend the comment, but then again there are many
things we could add a comment for.