`magit-display-process': fix timed process buffer popup
Commit e26c9c7 moved the timed process buffer popup logic from `magit-run*'
into `magit-display-process', but didn't take into account that it should
use `pop-to-buffer' instead of `display-buffer'.

When `magit-process-popup-time' was > 0 (as opposed to == 0), we *did*
use `display-buffer' instead of `pop-to-buffer', but since its docstring
explicitly states "Popup the process buffer...", *always* use `pop-to-buffer'
when we have a running process and `magit-process-popup-time' is >= 0.

Fixes #987.

Reported-by: David Glasser <glasser@davidglasser.net>
Signed-off-by: Pieter Praet <pieter@praet.org>