magit-log-longline-re: deal with three kinds of lines
There are up to three kinds of "log lines" that the `magit-log-{*}-re'
functions have to match.

* "actual log"   appearing in   long, oneline, others
* "graph-only"   appearing in   long, oneline
* "details"      appearing in   long

That might not be such a good idea - it certainly complicates writing
the regexps.  For now just make the regexps fully aware of that fact.

For "long log" that means that a line either contains a hash (and
possibly refs) **or** it has "something else".  For other log types that
would be the "message", but for "long log" that does not make much sense
- it is better to think of it as "the rest" after the (optional) graph.

In order not to complicate things we keep pretending this is "the
message" but we adjust the regexp to reflect the distinction.  By doing
so we prepare it for future changes.