magit-parse-log-line: stop misleading the reader
`magit-wash-log-line' is called repeatedly by `magit-wash-sequence'
until the latter returns nil or the end of the buffer or narrowed part
is reached.

`magit-wash-log-line' never returns nil, the explicit `t' at its end
makes sure of that.  And even without that it would never return nil.

So the reason `magit-wash-sequence' ever stops calling the washer is
that the end of the buffer or narrowed part is reached.  In practice
that means that all callers of `magit-wash-log' have to narrow to the
relevant portion before initiating the washing.

In other words `magit-parse-log-line' only ever gets to see lines that
are actual log lines.  It is therefor misleading that its body is
wrapped in a `when' form.