magit-wash-log-line: absorb magit-{parse,format}-log-line
Merge `magit-format-log-line' and `magit-parse-log-line' into
`magit-wash-log-line'.  `parse' was only used by `format', which in turn
was only used by `wash'.  Splitting the various subtasks into separate
functions did only complicate things.

In the process remove the `magit-log-line' struct which was only ever
used to pass an ordered collection of values from `parse' to `format'.
That was quite silly in itself, a simple would have been equally
powerful and much more compact.  This is just the tip of the iceberg.