magit-wash-log-line: make complete entry part of section
Function `magit-wash-log-line' turns each line that actually represents
a log entry into a section.  (It does not do so for graph-only lines).

When washing "oneline" logs that works find, but a "long" log entry
actually consists of all lines up to the next entry.  This commit
teaches `magit-wash-log-line' how to make the complete entry part of
the section.