magit-rev{,-range}-{to-git,describe}: don't raise an error if REV is nil
The helpers `magit-rev-to-git', `magit-rev-range-to-git',
`magit-rev-describe', and `magit-rev-range-describe' should not check
whether the REV or RANGE argument passed to them is nil and in that case
raise an error.  The REV or RANGE being nil is an error that should be
caught way before it reaches this function.  In fact I believe that this
error handling was added here because in the past `magit-read-rev' and
`magit-read-rev-range' did not raise an error on empty input.  The
latter is fixed now, so this kludge can be removed.