instantly convert "." to the value of magit-marked-commit
Until this commit `magit-rev-to-git' converted the string "." to the
value of `magit-marked-commit', and `magit-rev-describe' replaced it
with the string "mark".

There is only one place where such a "." string to represent
`magit-marked-commit' could come from: the user providing that string
when asked for a revision by `magit-read-rev'.  So instead of passing
that ominous value through many functions that have no understanding of
its meaning, instanly convert it to the value of `magit-marked-commit'.

The only (mis-)feature this loses is that `magit-rev-describe' now
inserts the actual rev or ref `magit-marked-commit' points to instead
of the, not so helpful, "mark" string.