magit-refresh-log-buffer: don't treat --all-match like --all
`--all' means: Pretend as if all the refs in refs/ are listed on the
command line as <commit>.

`--all-grep' means: Limit the commits output to ones that match all
given --grep, instead of ones that match at least one.

One has nothing to do with the other.  In the latter case the commits
are still limited by some revision argument.