magit-refresh-log-buffer: stop using magit-rev-range-describe
In `magit-refresh-log-buffer' stop using `magit-rev-range-describe'.
Currently revision ranges are passed around as cons cells, which is
wrong for several reasons.  Because this commit is only one of the
steps required to undo that design error, it does contain a kludge.

This changes the buffer headings for ranged/other logs from:

  Commits in RANDOM-NAME-FOR-REVISION
  Commits from RANDOM-NAME-FOR-REVISION-A to RANDOM-NAME-FOR-REVISION-B

to

  Commits in REVISION-AS-GIVEN-BY-USER
  Commits in REVISION-A-AS-GIVEN-BY-USER..REVISION-B-AS-GIVEN-BY-USER

That's one small step backward for prettiness, one giant leap forward
for correctness.  Currently Magit cannot tell A..B and A...B apart,
this, and other changes, will eventually make fixing this possible.
And displaying SOME-REF-MATCHING-REV isn't helpful when the user
explicitly inserted something like "HEAD~5".