magit-refresh-diff-buffer: stop using magit-rev-range-describe
In `magit-refresh-log-buffer' stop using `magit-rev-range-describe', by
replacing several design errors with temporary kludges.  The kludges are
necessary because the errors are not locally contained and over time
have propagated throughout `magit.el'.  One of the mistakes this commit
tries to address is that revision ranges are passed around as cons
cells, which is wrong for several reasons.

This changes the buffer headings for relative-to-worktree/range/stash
diffs from:

  Changes in RANDOM-NAME-FOR-REVISION
  Changes from RANDOM-NAME-FOR-REVISION-A to RANDOM-NAME-FOR-REVISION-B
  Changes from RANDOM-NAME-FOR-stash@{N}^2^ to stash@{N}

to

  Changes in REVISION-AS-GIVEN-BY-USER
  Changes in REVISION-A-AS-GIVEN-BY-USER..REVISION-B-AS-GIVEN-BY-USER
  Changes in stash@{N}^2^..stash@{N}

That's one small step backward for prettiness, one giant leap forward
for correctness.  Currently Magit cannot tell A..B and A...B apart,
this, and other changes, will eventually make fixing this possible.
And displaying SOME-REF-MATCHING-REV isn't helpful when the user
explicitly inserted something like "HEAD~5".  In case of stash diffs
the old pretty heading was just wrong and the new heading is very
ugly - ugly is better.