remove function magit-rev-range-to-git
Remove the function `magit-rev-range-to-git'.  Since that was the last
member of the section titled "worrisome to-git converters", that heading
can be removed too.  The previous commits have demonstrated that these
functions were not necessary (nothing of value (as in: not harmful) was
actually lost by removing them) and their existence did not actually
reduce code size (they, combined with their use, increased it by 60
lines).