don't store face section in a diffstat struct a third time
It is a common issue that the type of a section is stored in the section
struct twice, in slot TITLE and TYPE.  Sometimes one is a string and the
other a symbol, but there is little consistency which of the two is what.
And then these slots are also used as magic switches for only slightly
related functionality.  That's all pretty bad in itself, but...

For diffstat sections this was taken a step further.  The symbol
`diffstat' was additionally also stored as the car of the INFO slot.
And there even was a dedicated function to retrieve that information,
but it was never used.  Remove that function and don't cons `diffstat'
onto INFO, its "not necessary".