magit-wash-diffstat{,s}: use a symbol as the section "title"
The TITLE slot of section structs is not actually the "title".  It's
just another slot like TYPE used to store what kind of section it is,
except when it isn't (e.g. because it is nil, which appears to mean
"don't highlight the section".  So this is a kludge and doesn't have
to (actually cannot possibly) make sense.