merge magit-diffstats-ediff into magit-ediff
What we usually do when some command has to behave differently depending
on the type of the section at point, is to use `magit-section-action'.
Here we don't do that because it only has to behave differently on one
specific type, but conceptually we are doing the same.

Before this commit something completely different was done, resulting in
an unnecessary command `magit-diffstats-ediff', and an unnecessary keymap
`magit-diffstat-keymap' which was used as the keymap of a small portion
of the buffer (the diffstat).  We don't do that anywhere else, so stop
doing it here too.