magit-wash-diffstat: remove insufficient comment
Remove the comment "Don't decode pseudo filename." because it just
states what we (don't) do not why, and the why is to long for a comment
and also likely not what the initial reason was:

We not only don't decode the filename because it could be abbreviated
and that could cause problems, but because it is actually a good thing
to display the filename "as git would format it" somewhere.  That name
is only used for display, so that won't cause any problems when acting
on diffstat sections.