magit-goto-section: don't start showing details in long log
When moving between commits in the log buffer the commit buffer is
automatically updated to show the commit at point.  If the commit buffer
is not already displayed in the current frame then it is automatically
made current in another frame.

Previously this was done for both oneline and long logs.  Now it is only
done for oneline logs, because otherwise there would be little reason to
use the long log.  Since all the information displayed in long log but
not in oneline log is also displayed in the commit buffer, long log
until now did just waste space.

Of course it is still possible to explicitly display the commit buffer.
And once it is the current buffer in a window of the selected frame it
is also automatically updated, just like for oneline log.