soft require magit-log-edit
The old commit mode is now available as a separate package.  To make it
easier for its users soft require it.  If they have it installed it will
be loaded, causing some things defined in `magit.el' to be overridden.
If someone has it installed but doesn't want it to be loaded then he
should make up his mind.