magit-insert-untracked-files: make consistent and shorter
Use `git status' instead of `git ls-files' to get list of untracked
files, as recommended in the latter's manpage.  Reduce code size by
using `magit-with-section' instead of `magit-git-insert-section'.
The latter should only be used for complex washing.  (But in many
cases that isn't so yet.)

Always rely on `status.showUntrackedFiles'.  Previously that was only
used when its value was "no" and the "normal" (show untracked files
and directories) and "all" (also show individual files in untracked
directories) had to be set using `magit-status-verbose-untracked'.
That is highly confusing, either respect the git variable or don't.