magit-describe-item: move definition
Move the definition of `magit-describe-item' to the new "Maintenance
Tools" subsection of toplevel "Miscellaneous".  This new section will
soon contain more such things.

Rename the subsection of "Acting" where `magit-describe-item' from, from
"Describe" to "Kill".  The name previously was wrong because it also
contained and still does, the unrelated `magit-copy-item-as-kill'.

Now that it has been mostly cleaned up, we should probably lose the
subsections of the second "Acting" section, and rename it to something
like "Item Commands".  But that's another issue.