git-commit-auto-mode-enable: don't call on load and autoload
This is a somewhat controversial issue.  Some think that simply loading
or autoloading a library should prepare its modes for use.  Others like
me think that users should have to do that explicitly.  It might be a
bit annoying that an additional step is required to use the library and
its modes - it would be more annoying to have to do something *not* to
use it.

The argument "But why would they install it if they did not want to use
it?" does not apply it - Magit does force Melpa users to install it,
even though `magit-log-edit' could be used instead.  Magit has to ensure
one of the commit modes is installed and does so by making the preferred
one a dependency.  `package does not support optional dependencies or
even dependency alternatives.