magit-uninteresting-refs: always ignore topgit refs
Add the regexps that cause the topgit "top-base" refs to be ignored to
the definition of `magit-interesting-refs' instead of adding them only
when `magit-topgit-mode' is turned on.

I guess it could happen that someone does not use topgit and uses
"refs/top-bases" and "refs/remotes/REMOTE/top-bases" for something else,
but I think it that is rather unlikely.  If it does happen, that user
can just remove the respective regexps from the variable - which by the
way will get easier once it is a customizable option.