magit-list-interesting-refs: never trim refs
Always return  ((NAME . FULL-REF)...)
instead of     ((NAME . FULL-OR-ABBREVIATED-REF)...).

Previously the cdr in all but one case was the full ref.  The exception
was the most common case when REF began with "refs/heads/", in which
case that part was removed from the ref.

The old behavior was not helpful.  Sure, the (now) only caller needs the
refs to be abbreviated when they are branches, but that isn't an inherent
requirement - on the contrary that caller also has issues.