magit-format-ref: do not also filter
Previously `magit-format-ref' did return nil for refs other than
"refs/{head,remote,tag}/...".  Stop doing that, formatting should
not be overloaded to also filter.

As a result some refs are now displayed in logs, that previously were
not.  For some users this will be a regression, for others it will be a
welcome change.

Once `magit-uninteresting-refs' is an option, it will be easy for users
in the first group to, "fix" this change.