magit-wip-mode: remove mode
Remove `magit-wip-mode' which was used to give "refs/wip/BRANCH" a
special appearance and nothing else; and obviously only in buffers where
it was actually enabled.  Don't confuse it with `magit-wip-save-mode'.

This commit removes the special appearance temporarily, but in a few
commits that will be restored.  Unfortunately this use of the existing
"ref formatting" code prevents changes to that code, which forces us to
do it like this.

The special appearance won't be restored exactly as it is now: it will
drop the "(WIP) " prefix.  No other "labels" for other kinds of refs use
such a prefix, and it also isn't necessary, because we already use a
different face to give it distinct look.  Also in a few commits, the
ability for users to add such a prefix using the Custom interface will
be added for all ref namespaces, not just this particular one.  But we
won't use it for any refs by default.

There will also soon be a Custom option that allows hiding certain refs.

By removing this mode we lose the ability to only give "refs/wip/BRANCH"
a special appearance in certain buffers instead of all.  However it soon
will be possible again to give these refs the fallback appearance by
customizing an option.

I don't really think this is a loose at all: if such refs exist then
that most likely means that the user makes use of `magit-wip' and/or
`git wip'; and would like them to have a special appearance everywhere
or nowhere.  If he doesn't use these tools but uses the namespace for
some other reason and doesn't want them to look any different, then
again he can just customize the option.

If the user does use `*wip' in some repositories and wants them to have
a special appearance there, but also uses refs in the same namespace for
other purposes in other repositories, then he is doing it wrong and we
shouldn't keep the mode just so that he can keep on doing that.  Just
saying, I don't expect anyone actually does that.