magit-format-ref-label: don't abuse it for filtering
The last few commits were already preparing for this; here we just make
it final: Do not abuse `magit-format-ref-label', which is supposed to
put nice colors on refs and possibly change their text for display, to
also _filter out_ certain refs.  Performing an additional, completely
unrelated task, made this function unreadable, and required the addition
of several functions and a hook variable.  Remove those that remained
until now.

There is a mechanism dedicated to the filtering of refs:
`magit-list-interesting-refs'.  (That also has issues but that is
a different story).

This isn't perfect yet.  E.g. the involved faces are still named
`magit-log-head-label-{*}'; which isn't good because this, now that
sanity has been restored, could be used elsewhere (but isn't yet).  Also
these faces don't have anything to do with the HEAD, so it makes little
sense that their names contain "-head-".