magit-refs-namespaces: improve rule for patches refs
Previously the label text for a ref "refs/patches/SOMETHING" was
"patches/SOMETHING".  We don't do that for any other ref namespace, and
I think this was only done this way because of limitations with the old
code and not because someone actually wanted the prefix to be displayed.
Regardless this makes this namespace consistent with the others, and if
someone wants to keep this prefix, or all prefixes, then she can
customize this option.