magit-reset-head{,-hard}: use magit-guess-branch
In `magit-reset-head' and `magit-reset-head-hard' use
`magit-guess-branch' to determine the default branch for completion.

Previously `magit-default-rev' was used but in many cases that ignored a
potential default.  Also the explicit fallback "HEAD" (or "HEAD^") is
now much more likely to be used.  Previously the "previous branch" was
used as fallback (in `magit-default-rev') and that almost never is
nil (and makes much less sense).