magit-tag: use magit-guess-branch
In `magit-tag' use `magit-guess-branch' to determine the default offered
during completion of the ref/rev to be tagged.

Previously `magit-default-rev' was used but in many cases that ignored a
potential default.  Also the explicit fallback "HEAD" is now much more
likely to be used.  Previously the "previous branch" was used as
fallback (in `magit-default-rev') and that almost never is nil (and
makes much less sense).