magit-list-interesting-refs: UNINTERESTING may no longer be a function
The UNINTERESTING argument of `magit-list-interesting-refs' now has to
be a list of regexps, or nil.  A function is no longer accepted.  That
was only used by one caller and that used it to filter out refs that it
should not have filtered out (fixed in previous commit).

Since this function will be removed soon anyway, don't just adjust the
doc-string, remove it instead.