magit-list-interesting-refs: support branch for UNINTERESTING
When argument UNINTERESTING is a string (instead of nil or a list of
regexps) then treat it as a branch name and arrange for it as well as
the defaults in `magit-uninteresting-refs' to be ignored.

The definition of `magit-uninteresting-refs' has to be moved to avoid
a compiler warning about it not being defined.