remove magit--is-branch-at-point-remote
Remove `magit--is-branch-at-point-remote'.  This function was still used
by `magit-change-what-branch-tracks' which due to this change no longer
performs the respective test.  That's okay, there are quite a few (more
important) things this function fails to do, and fixing that will be
easier if the function being removed here is not used (and no temporary
substitute is hacked in).