magit-delete-branch: don't prepend refs/heads when push deleting
In `magit-delete-branch' when deleting a remote branch by pushing
"nothing" to it do not prepend "refs/heads/".  That almost always
is the full name of what we want to delete, but it not necessary
to use the full name.  On the other hand pushing to the partial
symbolic name works even with non-standard remote.<remote>.fetch.