move definition of magit-map-magit-buffers
Move the definition of function `magit-map-magit-buffers' into a new
subsection of "Magit Mode", named "Mode Utilities".  That isn't optimal
because its the only function in that new section, but that might change
over time.  Randomly sticking it into one of the section that happens to
contain a function that uses it, is what we did before.