move "Refresh Machinery" section to toplevel
Move the "Refresh Machinery" section to the toplevel.  Previously it
was a subsection of "Magit Mode".  The main reason it was merely a
subsection for a long time, is that moving it to the toplevel did
require some code to be moved around (see previous commits).

The refresh machinery is only loosely coupled with the code related
to the code in "Magit Mode".  This move just officializes this, and
was prone to happen eventually.  Since I have plans to improve both
the "mode" and the "refresh" logic, that moment has come now.