move definitions of magit-refresh-{function,args}
Move the definitions of the variables `magit-refresh-functions' and
`magit-refresh-args' to another subsection of "Magit Mode".  Previously
they were located inside the subsection "Hook" which somewhat made sense
because they are used in a hook-like fashion.

However the other functions in that section of "Magit Mode" are hook
functions that are necessary because of limitations in the section
code (or used by them).  This limitations will eventually be lifted
making these kludges unnecessary.

The current heading of the new section, "Initialize Mode", doesn't do
the moved functions full justice and is subject to change.