merge three subsections of "Magit Mode"
Merge the subsections "Initialize Mode", "Display Buffer", and "Refresh
Buffer" into a new subsection "Mode Foundation".  The old distinctions
did make sense, but after past cleanup it is no longer *necessary* to
keep them separated to improve the odds of understanding them.  The
functions and variables are closely related, and interconnected, so
having them in the same section makes sense.