remove function magit-find-status-buffer
Remove function `magit-find-status-buffer' and just use generic
`magit-find-buffer'.  The latter will be renamed and extended in
subsequent commits and then more often used for other modes too;
it which point it won't makes sense anymore, to have a specialized
function for just one of the modes (and adding one for all modes
certainly isn't an option).