move magit-mode-quit-window to end of its section
Move the function `magit-mode-quit-window' from the middle to the end of
section "Mode Foundation".  It is closely related to the other functions
in this section but is different in that does not "create" a mode buffer
but makes it go away.  The end seems like a suitable place for that.