add option magit-commit-ask-to-stage
Add option `magit-commit-ask-to-stage'.  If it is non-nil and nothing
is staged then `magit-commit-assert' expands the unstaged changes.

This partially restores a similar feature that was available before the
switch to `git-commit-mode', and also extends it.  It is now less
flexible in that it is a simple boolean.  On the other hand it shows
the unstaged changes to the user, making it easier for him to make up
his mind.