add new command magit-diff-item
Add the new command `magit-diff-item' and bind it to "=" in
`magit-status-mode-map'.

This makes it easier to generate per-item diff buffers for individual
items in the status buffer, and with the "=" binding it is more
consistent with both VC and PCL-CVS.

Ideally the diff buffers would have the same interface functionality as
an expanded changes section in the main status buffer such that
operations such as hunk staging could be performed directly in
them (thus allowing them to be in a different window or frame).  However
this added functionality is as yet unimplemented.