magit-wash-log-line: use magit-format-ref-labels
This has to unfortunate but acceptable consequence that when there are
only "uninteresting" refs then, when showing a "long" log the revision
hash is abbreviated just if there were at least one "interesting" ref.
In the latter case we do so to make room for the displaying the refs,
in the former case we don't need to make room, but now do so anyway.
However that doesn't really matter.