magit-bisect-reset: no error if nothing to reset
Running `git bisect reset` when there is no running or finished bisect
does nothing (no error).  That's good enough for use, let's not perform
a needless test and raise an error if it fails.

That will also make it easier later on.  We start to benefit in this
commit already by removing the "reset" special case in
`magit--bisect-cmd'.